% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoverTree.R
\name{add.data.CoverTree}
\alias{add.data.CoverTree}
\title{Add Data}
\usage{
add.data(ct,data)
}
\arguments{
\item{data}{a data.frame containing the data to be added to the cover tree}

\item{ct}{The CoverTree to be displayed}
}
\value{
n/a
}
\description{
Inserts additional data into the cover tree.
}
\details{
The column names in the input data frame must exactly match those in the data used to create
the CoverTree object.

Redundant data will not be added to the cover tree.
}
