% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoverTree.R
\name{as.dendrogram.CoverTree}
\alias{as.dendrogram.CoverTree}
\title{Dendrogram data}
\usage{
\code{> dend <- as.dendrogram(ct)}
\code{> plot(dend) }
}
\arguments{
\item{ct}{The CoverTree to be converted to be converted}
}
\value{
data structure that can be plotted as a dendrogram
}
\description{
Dendrogram data
}
\details{
Returns a data structure that can be plotted as hclust dendrograms
}
