% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoverTree.R
\name{as.hclust.CoverTree}
\alias{as.hclust.CoverTree}
\title{Dendragram data}
\usage{
\code{> hc <- as.hclust(ct)}
\code{> plot(hc) }
}
\arguments{
\item{ct}{The CoverTree to be converted to be converted}
}
\value{
data structure that can be plotted as a dendragram
}
\description{
Dendragram data
}
\details{
Returns a data structure that can be plotted as hclust dendragrams
}
