% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoverTree.R
\name{as.nodes.CoverTree}
\alias{as.nodes.CoverTree}
\title{CoverTree node list}
\usage{
\method{as.nodes}{CoverTree}(self)
}
\arguments{
\item{ct}{The CoverTree containing the nodes}
}
\value{
data.frame with columns: parent, level, distance
}
\description{
CoverTree node list
}
\details{
Returns a data.frame with the following columns:
\itemize{
 \item{\strong{level}: level of the current node}
 \item{\strong{parent}: index of the row in the data which is the parent node of the current node}
 \item{\strong{distance}: distance from the current node to the parent node}
 }
 The parent node and distance to parent are both set to 0 for the root node
}
