% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoverTree.R
\name{CoverTree}
\alias{CoverTree}
\title{Cover Tree}
\usage{
ct <- CoverTree(data,dist.func[,param])
}
\arguments{
\item{data}{a data.frame containing the data to be covered}

\item{dist.func}{a function that computes the distance between 2 entry rows in the data}
}
\value{
An initialized CoverTree as described above
}
\description{
Implementation of a Cover Tree algorithm for a specified set of data
}
\details{
The implementation here is based on the rules and insertion algorithm defined in
\href{http://hunch.net/~jl/projects/cover_tree/paper/paper.pdf}{Alina Beygelzimer, Sham Kakade, and John Langford: Cover Trees for Nearest Neighbor. (Web)}
The current implementation, however, negates the levels outlined in that algorithm.
\itemize{
\item{All nodes in level i are implicitly in level i+1}
\item{All pairs of nodes in level i have a separation of more than 1/(2^i)}
\item{All nodes in level i+1 have exactly one parent in level i that is no farther away than 1/(2^i)}
}

The provided data must be a data.frame with at least two rows of unique data.
For the purpose of the CoverTree algorithm, unique data means having a non-zero separation.
\strong{Redundant data will not be added to the cover tree}.

The provided distance function \strong{must} take two rows from the input data as
arguments and must return a distance.   The returned distance must:
\itemize{
\item{be non-negative}
\item{be symmetric: \emph{dist(a,b) = dist(b,a)}}
\item{obey the triangle rule: \emph{dist(a,b) + dist(b,c) >= dist(a,c)}}
}

The provided distance function, \strong{may} take an optional third parameter, which may be of any valid R data type.
If it does, this parameter must be provided as the third argument to cover tree constructor.
This parameter will be passed to every invocation of the distance function.

The columns defined in the data.frame are irrelevant to the CoverTree algorithm, but must be
consistent with the distance function.
}
