% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoverTreeNode.R
\name{CoverTreeNode}
\alias{CoverTreeNode}
\title{Cover Tree Node}
\usage{
CoverTreeNode(row, data, parent = NULL, distance = NULL)
}
\arguments{
\item{row}{index associated with the data}

\item{data}{A valid R object containing the data represented by the node}

\item{parent}{The parent node for the one being constructed (n/a for root node)}

\item{distance}{Either a numeric value or a distance function [ f(data,data) -> numeric ]}
}
\value{
An initialized CoverTreeNode, fully connected as a child of 'parent'
}
\description{
Constructs a cover tree node.  This constructor should proably never
be called directly.   Rather it should be called by CoverTree during
its construction and subsequent updates.
}
\seealso{
\link{CoverTree}
}
