% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoverTree.R
\name{plot.CoverTreeDendrogram}
\alias{plot.CoverTreeDendrogram}
\title{Plot CoverTree Dendrogram}
\usage{
\code{> dend <- as.dendrogram(ct) }
\code{> dend$labels <- x }
\code{> plot(dend) }
}
\arguments{
\item{dend}{A cover tree dendrogram object}
}
\value{
n/a
}
\description{
Plot CoverTree Dendrogram
}
\details{
Returns a data structure that can be plotted as hclust dendrograms

Note that the CoverTreeDendrogram does \strong{not} provide data labels
If you wish to have the nodes labeled, you must append a labels dimension
to the CoverTreeDendrogram object as shown above

This plot function takes all of the same optional arguments taken by
plot.hclust.  See the link below for additional info.
}
\seealso{
\link{as.dendrogram.CoverTree}

\link{plot.hclust}
}
