% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoverTree.R
\name{print.CoverTree}
\alias{print.CoverTree}
\title{Print a CoverTree}
\usage{
\code{> print(ct)}
\emph{or}
\code{> ct}
}
\arguments{
\item{ct}{The CoverTree to be displayed}
}
\value{
n/a
}
\description{
Print a CoverTree
}
